Tests/LibWeb/Text/expected/hit_testing/inert-svg-toggle.txt

hit before inert: art
hit with inert: body
hit after inert removed: art
